Organize and cleanup components in case module
