+++ client/app/src/app/case/case.module.ts
@@ -1,10 +1,10 @@
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CaseComponent } from './case/case.component';
-import { EventComponent } from './event/event.component';
-//import { FormComponent } from './form/form.component';
-import { CasesComponent } from './cases/cases.component';
-import { NewCaseComponent } from './new-case/new-case.component';
+import { CaseComponent } from './components/case/case.component';
+import { EventComponent } from './components/event/event.component';
+//import { FormComponent } from './components/form/form.component';
+import { CasesComponent } from './components/cases/cases.component';
+import { NewCaseComponent } from './components/new-case/new-case.component';
 import { CaseRoutingModule } from './case-routing.module';
 
 @NgModule({
