+++ client/app/src/app/case/classes/case-event-definition.class.ts
@@ -1,18 +1,22 @@
 import UUID from 'uuid/v4';
 import { EventForm } from './event-form.class'
+import { EventFormDefinition } from './event-form-definition.class'
 
 class CaseEventDefinition {
   id:string;
   name:string;
   description:string;
   repeatable:boolean = false;
-  forms:Array<EventForm> = []
-  constructor(data) {
+  eventFormDefinitions:Array<EventFormDefinition> = []
+  constructor(init:CaseEventDefinition) {
+    Object.assign(this, init);
+    /*
     this.id = data.id 
     this.name = data.name
     this.description = data.description
     this.repeatable = data.repeatable
     this.forms = data.forms.map(formData => new EventForm(formData))
+    */
   }
 }
