+++ client/app/src/app/case/classes/event-form-definition.class.ts
@@ -1,0 +1,10 @@
+export class EventFormDefinition {
+  id:string
+  formId:string
+  name:string
+  required:boolean
+  repeatable:boolean
+  public constructor(init:EventFormDefinition) {
+    Object.assign(this, init);
+  }
+}
