+++ client/app/src/app/case/components/event/event.component.spec.ts
