--- client/app/src/app/case/event/event.component.html
@@ -1,14 +1,0 @@
-<h1>{{caseEvent ? caseEvent.name : ''}}</h1>
-<!--
-<p>
-  {{event.description}}
-</p>
-<ul>
-  <li *ngFor="let form of event.forms">
-    <a href="#/case/{{case._id}}/event/{{event._id}}/{{form.formId}}">
-      <span *ngIf="form.required === true">*</span>
-      {{form.title}}
-    </a>
-  </li>
-</ul>
--->
