--- client/app/src/app/case/form/form.component.html
@@ -1,3 +1,0 @@
-<p>
-  form works!
-</p>
